requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
pandas>=2.0.0
requests>=2.28.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0
